Use error_handling decorator only for Micka methods, which communicate with Micka
