+++ src/layman/common/micka/__init__.py
@@ -1,0 +1,12 @@
+NAMESPACES = {
+    'csw': 'http://www.opengis.net/cat/csw/2.0.2',
+    'ows': 'http://www.opengis.net/ows/1.1',
+    'dc': 'http://purl.org/dc/elements/1.1/',
+    'gco': 'http://www.isotc211.org/2005/gco',
+    'gmd': 'http://www.isotc211.org/2005/gmd',
+    'gmx': 'http://www.isotc211.org/2005/gmx',
+    'xlink': 'http://www.w3.org/1999/xlink',
+    'srv': 'http://www.isotc211.org/2005/srv',
+    'soap': 'http://www.w3.org/2003/05/soap-envelope',
+    'hs': 'http://www.hsrs.cz/micka',
+}
