+++ src/layman/common/micka/util.py
@@ -9,22 +9,12 @@ from owslib.csw import CatalogueServiceWeb
 from owslib.util import nspath_eval
 from lxml import etree as ET
 
-from layman import settings, LaymanError, authz
+from layman import settings, authz
 from layman.common.metadata import PROPERTIES as COMMON_PROPERTIES
 from layman.util import get_publication_info
+from . import NAMESPACES
+from .requests import base_insert, csw_delete, fill_template_as_str
 
-NAMESPACES = {
-    'csw': 'http://www.opengis.net/cat/csw/2.0.2',
-    'ows': 'http://www.opengis.net/ows/1.1',
-    'dc': 'http://purl.org/dc/elements/1.1/',
-    'gco': 'http://www.isotc211.org/2005/gco',
-    'gmd': 'http://www.isotc211.org/2005/gmd',
-    'gmx': 'http://www.isotc211.org/2005/gmx',
-    'xlink': 'http://www.w3.org/1999/xlink',
-    'srv': 'http://www.isotc211.org/2005/srv',
-    'soap': 'http://www.w3.org/2003/05/soap-envelope',
-    'hs': 'http://www.hsrs.cz/micka',
-}
 for k, v in NAMESPACES.items():
     ET.register_namespace(k, v)
 
@@ -185,13 +175,6 @@ def fill_template(template_path, template_values):
     return file_object
 
 
-def fill_template_as_str(template_path, template_values):
-    with open(template_path, 'r') as template_file:
-        template_str = template_file.read()
-    xml_str = template_str.format(**template_values)
-    return xml_str
-
-
 def fill_template_as_el(template_path, template_values):
     xml_str = fill_template_as_str(template_path, template_values)
     parser = ET.XMLParser(remove_blank_text=True)
@@ -228,39 +211,6 @@ def create_csw():
     opts['skip_caps'] = True
     csw = CatalogueServiceWeb(settings.CSW_URL, **opts) if settings.CSW_URL is not None else None
     return csw
-
-
-def is_record_exists_exception(root_el):
-    return len(root_el) == 1 and \
-        root_el[0].tag == nspath_eval('ows:Exception', NAMESPACES) and \
-        "exceptionCode" in root_el[0].attrib and \
-        root_el[0].attrib["exceptionCode"] == 'TransactionFailed' and \
-        len(root_el[0]) == 1 and \
-        root_el[0][0].tag == nspath_eval('ows:ExceptionText', NAMESPACES) and \
-        root_el[0][0].text.startswith('Record exists')
-
-
-def base_insert(xml_str):
-    # print(f"Micka insert=\n{xml_str}")
-    response = requests.post(settings.CSW_URL,
-                             auth=settings.CSW_BASIC_AUTHN,
-                             data=xml_str.encode('utf-8'),
-                             timeout=settings.DEFAULT_CONNECTION_TIMEOUT,)
-    # print(f"Micka insert response=\n{r.text}")
-    response.raise_for_status()
-    root_el = ET.fromstring(response.content)
-
-    if root_el.tag == nspath_eval('ows:ExceptionReport', NAMESPACES):
-        if is_record_exists_exception(root_el):
-            raise LaymanError(36, data={
-                'exception_code': root_el[0].attrib["exceptionCode"],
-                'locator': root_el[0].attrib["locator"],
-                'text': root_el[0][0].text,
-            })
-        raise LaymanError(37, data={
-            'response': response.text
-        })
-    return root_el, response
 
 
 def csw_insert(template_values):
@@ -304,67 +254,6 @@ def soap_insert_record(record, is_public):
 def soap_insert_record_from_template(template_path, prop_values, metadata_properties, is_public):
     record = fill_xml_template_as_pretty_str(template_path, prop_values, metadata_properties)
     return soap_insert_record(record, is_public)
-
-
-def csw_update(template_values, timeout=None):
-    timeout = timeout or settings.DEFAULT_CONNECTION_TIMEOUT
-    template_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'csw-update-template.xml')
-    xml_str = fill_template_as_str(template_path, template_values)
-    # print(f"CSW update request=\n{xml_str}")
-    response = requests.post(settings.CSW_URL,
-                             auth=settings.CSW_BASIC_AUTHN,
-                             data=xml_str.encode('utf-8'),
-                             timeout=timeout,
-                             )
-    # print(f"CSW update response=\n{r.text}")
-    response.raise_for_status()
-    root_el = ET.fromstring(response.content)
-
-    if root_el.tag == nspath_eval('ows:ExceptionReport', NAMESPACES):
-        if is_record_does_not_exist_exception(root_el):
-            raise LaymanError(39, data={
-                'response': response.text
-            })
-        raise LaymanError(37, data={
-            'response': response.text
-        })
-    assert root_el.tag == nspath_eval('csw:TransactionResponse', NAMESPACES), response.content
-    assert root_el.find(nspath_eval('csw:TransactionSummary/csw:totalUpdated', NAMESPACES)).text == "1", response.content
-
-
-def is_record_does_not_exist_exception(root_el):
-    return len(root_el) == 1 and \
-        root_el[0].tag == nspath_eval('ows:Exception', NAMESPACES) and \
-        "exceptionCode" in root_el[0].attrib and \
-        root_el[0].attrib["exceptionCode"] == 'TransactionFailed' and \
-        len(root_el[0]) == 0 and \
-        root_el[0].text is None
-
-
-def csw_delete(muuid):
-    template_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'csw-delete-template.xml')
-    template_values = {
-        'muuid': muuid
-    }
-    xml_str = fill_template_as_str(template_path, template_values)
-    # print(f"CSW delete request=\n{xml_str}")
-    response = requests.post(settings.CSW_URL,
-                             auth=settings.CSW_BASIC_AUTHN,
-                             data=xml_str.encode('utf-8'),
-                             timeout=settings.DEFAULT_CONNECTION_TIMEOUT,
-                             )
-    # print(f"CSW delete response=\n{r.text}")
-    response.raise_for_status()
-    root_el = ET.fromstring(response.content)
-
-    if root_el.tag == nspath_eval('ows:ExceptionReport', NAMESPACES):
-        if is_record_does_not_exist_exception(root_el):
-            return
-        raise LaymanError(37, data={
-            'response': response.text
-        })
-    assert root_el.tag == nspath_eval('csw:TransactionResponse', NAMESPACES), response.content
-    assert root_el.find(nspath_eval('csw:TransactionSummary/csw:totalDeleted', NAMESPACES)).text == "1", response.content
 
 
 def parse_md_properties(file_obj, property_names, publ_properties):
